requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
click==8.1.8
constantly==23.10.4
crispy-tailwind==1.0.3
cryptography==43.0.3
daphne==4.1.2
Django==5.1.1
django-crispy-forms==2.3
django-tailwind==3.8.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
redis==5.1.1
service-identity==24.1.0
setuptools==75.2.0
six==1.17.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
uvicorn==0.34.0
zope.interface==7.1.0
